create a new js file to hold the functionality for interactivity
